django/requirements.txt
sgiref==3.5.2
Django==4.0.6
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
sqlparse==0.4.2
tzdata==2022.1
djangorestframework==3.13.1
django-filter==22.1
markdown==3.4.1